Moved login view to views folder
