+++ app/src/router.js
@@ -5,8 +5,8 @@ import Popup from './views/Popup.vue'
 import Confirm from './views/Confirm.vue'
 import ProviderChange from './views/ProviderChange.vue'
 import Wallet from './views/Wallet.vue'
+import Login from './views/Login.vue'
 import WalletHome from './containers/WalletHome.vue'
-import Login from './containers/Login.vue'
 import WalletHistory from './containers/WalletHistory.vue'
 import WalletSettings from './containers/WalletSettings.vue'
 import WalletAccounts from './containers/WalletAccounts.vue'
